crowd_app/requirements.txt
Django==5.2
djangorestframework==3.15.1
django-cors-headers==4.4.0
drf-yasg==1.21.5
Pillow==10.3.0
celery==5.3.6
redis==5.0.1
django-redis==5.3.0
django-cleanup==8.0.0